+++ R/util.R
@@ -19,10 +19,10 @@ getAcovs <- function(possible_patterns, method, test, design){
 } 
 
 weighted_mat <- function(mat, row_loc, which_not_answered, P = rep(1, length(row_loc))){
-    if(is.list(P)) for(i in 1L:length(P)) mat[[i]] <- mat[[i]] * P[[i]]
-    else for(i in 1L:length(P)) mat[[i]] <- mat[[i]] * P[i]
+    if(is.list(P)) for(i in seq_len(length(P))) mat[[i]] <- mat[[i]] * P[[i]]
+    else for(i in seq_len(length(P))) mat[[i]] <- mat[[i]] * P[i]
     mat2 <- vector('list', length(unique(row_loc)))
-    for(i in 1L:length(mat2)){
+    for(i in seq_len(length(mat2))){
         pick <- which(row_loc == which_not_answered[i])
         tmp <- mat[pick]
         for(j in 2L:length(pick))
@@ -106,7 +106,7 @@ integrate.xy <- function(x,fx, a,b, use.spline = TRUE, xtol = 2e-8)
     r/2
 }
 
-buildShinyElements <- function(questions, itemnames, customTypes){
+buildShinyElements <- function(questions, itemnames, customTypes, choiceNames, choiceValues){
     J <- length(questions$Type)
     if(!all(sapply(questions[names(questions) != 'Rendered_Question'], is.character))) 
         stop('Only character classes are supported in questions input', call.=FALSE)
@@ -124,15 +124,23 @@ buildShinyElements <- function(questions, itemnames, customTypes){
     choices <- data.frame(questions[grepl('Option', names)], stringsAsFactors = FALSE)
     choices_list <- vector('list', J)
     names(choices) <- NULL
-    for(i in 1L:length(Qs)){
+    for(i in seq_len(length(Qs))){
         if(Type[i] == 'radio'){
-            cs <- choices[i, !is.na(choices[i, ])]
-            choices_list[[i]] <- cs
+            cNs <- cVs <- cs <- NULL 
+            if(length(choiceNames[[i]]) && is.list(choiceNames[[i]])){
+                cNs <- choiceNames[[i]]
+                cVs <- choiceValues[[i]]
+                choices_list[[i]] <- as.character(cVs)
+            } else {
+                cs <- choices[i, !is.na(choices[i, ])]
+                choices_list[[i]] <- cs
+            }
             inline <- if(is.null(questions$inline[i])) FALSE else as.logical(questions$inline[i])
             width <- questions$width[i]
             Qs[[i]] <- radioButtons(inputId = itemnames[i], label='',
                                     inline = inline, width = width,
-                                    choices = cs, selected = '')
+                                    choices = cs, selected = '', 
+                                    choiceNames=cNs, choiceValues=cVs)
         } else if(Type[i] == 'select'){
             cs <- c('', choices[i, !is.na(choices[i, ])])
             choices_list[[i]] <- cs
@@ -156,6 +164,8 @@ buildShinyElements <- function(questions, itemnames, customTypes){
                                  width=width, height=height, placeholder=placeholder,
                                  cols=cols, rows=rows, resize=resize)
         } else if(Type[i] == 'slider'){
+            if(is.null(questions$min) || is.null(questions$max) || is.null(questions$step))
+                stop('slider Type requires a min, max, and step column element in the df object', call.=FALSE)
             VALUE <- as.numeric(ifelse(is.null(questions$value[i]), questions$min[i], questions$value[i]))
             MIN <- as.numeric(questions$min[i])
             MAX <- as.numeric(questions$max[i])
@@ -170,12 +180,19 @@ buildShinyElements <- function(questions, itemnames, customTypes){
                                    ticks = TICKS, pre = PRE, post = POST)
             choices_list[[i]] <- as.character(seq(from=MIN, to=MAX, by=STEP))
         } else if(Type[i] == 'checkbox'){
-            cs <- choices[i, !is.na(choices[i, ])]
-            choices_list[[i]] <- cs
+            cNs <- cVs <- cs <- NULL 
+            if(length(choiceNames[[i]]) && is.list(choiceNames[[i]])){
+                cNs <- choiceNames[[i]]
+                cVs <- choiceValues[[i]]
+                choices_list[[i]] <- as.character(cVs)
+            } else {
+                cs <- choices[i, !is.na(choices[i, ])]
+                choices_list[[i]] <- cs
+            }
             width <- questions$width[i]
             inline <- if(is.null(questions$inline[i])) FALSE else as.logical(questions$inline[i])
             Qs[[i]] <- checkboxGroupInput(inputId = itemnames[i], label='', choices = cs,
-                                          width=width, inline=inline)
+                                          width=width, inline=inline, choiceNames=cNs, choiceValues=cVs)
         } else if(Type[i] %in% names(customTypes)){
             nm <- names(customTypes)[names(customTypes) == Type[i]]
             df_row <- as.data.frame(lapply(questions, function(x, ind) x[[ind]], ind=i),